Added tests for versioned objects
* Expire current version
* Expire non current versions

Had to introduce minio-go client because this client
provides a possibility to set the modified date of
objects into the past. The rules for non current expiration
require objects to have modified dates in the past.

All other tests could be implemented by just creating
a lifecycle configuration with a date in the past.