Revert changes to trust a chain of TLS certificates
This breaks our current test setup. We only ever use self-signed
certificates. This used to work, but broke with commit
2f549cb9f0e6abc5c2f91ac708c239677cff5791 when MinIO decided to import an
entire chain of certificates.

Interestingly enough, the code works also for self-signed certificates
**if** one uses the *.crt file ending instead of *.pem. The way MinIO
implemented it with *.pem files, `update-ca-certificates` will not
import a single file as a new root CA.

We discussed if this is a change that could be upstreamed, but we don't
have a test setup to verify if changing the file ending doesn't break
anything when chains of certificates are involved.